Fixes after comment on PR
